+++ packages/components-js/tests/e2e/specs/tabs-bar.e2e.ts
@@ -42,9 +42,8 @@ describe('tabs-bar', () => {
   const clickHandlerScript = `
     <script>
       const tabsBar = document.querySelector('p-tabs-bar')
-      tabsBar.addEventListener('tabChange', (tabChangeEvent) => {
-          const { activeTabIndex } = tabChangeEvent.detail;
-          tabChangeEvent.target.setAttribute('active-tab-index', activeTabIndex);
+      tabsBar.addEventListener('tabChange', (e) => {
+          e.target.activeTabIndex = e.detail.activeTabIndex;
       });
     </script>`;
   const initTabsBar = async (opts?: InitOptions) => {
@@ -697,7 +696,7 @@ describe('tabs-bar', () => {
 
     it('should trigger event on button click', async () => {
       await initTabsBar({ amount: 3, activeTabIndex: 1 });
-      const host = await selectNode(page, 'p-tabs-bar');
+      const host = await getHost();
       const [firstButton, secondButton, thirdButton] = await getAllButtons();
       let eventCounter = 0;
       await addEventListener(host, 'tabChange', () => eventCounter++);