+++ packages/components-react/src/pages/index.ts
@@ -1,3 +1,4 @@
+export * from './Accordion';
 export * from './Banner';
 export * from './Button';
 export * from './ButtonGroup';