+++ packages/storefront/src/utils/formatting.ts
@@ -1,6 +1,6 @@
-import { camelCase, pascalCase } from 'change-case';
+import { camelCase, paramCase, pascalCase } from 'change-case';
 import { Theme } from '@/models';
-import { getComponentMeta } from '@porsche-design-system/shared';
+import { getComponentMeta, TagName } from '@porsche-design-system/shared';
 
 export const cleanMarkup = (markup: string): string =>
   markup
@@ -13,8 +13,10 @@ export const patchThemeIntoMarkup = (markup: string, theme: Theme): string =>
   theme === 'dark'
     ? markup
         // add dark theme attribute if component supports it
-        .replace(/(<p-[\w-]+)/g, (m, $tag) => {
-          return getComponentMeta($tag.replace('<', ''))?.isThemeable ? `${$tag} theme="dark"` : $tag;
+        .replace(/(<[pP][\w-]+)/g, (m, $tag) => {
+          return getComponentMeta(paramCase($tag.replace('<', '')) as TagName)?.isThemeable
+            ? `${$tag} theme="dark"`
+            : $tag;
         })
     : markup;