+++ packages/components/CHANGELOG.md
@@ -9,7 +9,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### [Unreleased]
 
 #### Added
-- `Checkbox Wrapper`, `Radiobutton Wrapper`, `Select Wrapper`, `Text Field Wrapper` and `Textarea Wrapper` now reflect changes of the `required` attribute on their child component
+- `Checkbox Wrapper`, `Radio Button Wrapper`, `Select Wrapper`, `Textarea Wrapper` and `Text Field Wrapper` now reflect changes of the `required` attribute on their child component
 
 #### Changed
 - Removed initial delay of `Banner`